:recycle: Fix "getNext" navigation function.
Use better typing in relevantSections
Use standard find methods instead of complex loop.

Note, it would probably be better to have the "getNext" function return the next section so that the "Agree and Next" button could be disabled if the next section was disabled.
I opted for a "loop around" pattern instead of the "just go nowhere and look broken approach".